pyproject.toml
[project]
name = "side-stacker-game"
version = "0.1.0"
description = ""
authors = [{ name = "Nicolas Epstein", email = "nilueps@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry]
packages = [{ include = "ssg", from = "src" }]

[tool.poetry.scripts]
start = "ssg:run"
init_db = "ssg:init_db"
train = "ssg.ai.training.train:main"

[tool.poetry.dependencies]
quart = ">=0.20.0,<0.21.0"
torch = { version = "^2.7", source = "pytorch-cpu", markers = "extra=='cpu' and extra!='cuda'" }

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"
numpy = "^2.2.6"

[tool.poetry.extras]
cpu = ["torch", "torchvision"]

[[tool.poetry.source]]
name = "pytorch-cpu"
priority = "explicit"
url = "https://download.pytorch.org/whl/cpu"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.basedpyright]
include = ["src"]
reportAny = "none"
reportExplicitAny = "none"
reportUnusedCallResult = "none"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
